Extract parent-side persona interaction tools from MCP server.ts
Move list_personas, dispatch_launch_persona, dispatch_get_feedback,
dispatch_resolve_feedback, and dispatch_submit_resolution into a new
persona-interaction-tools.ts module. Also moves buildLaunchPersonaResponseText.
Reduces server.ts from 1377 to 1073 lines.
